+++ programmable_video_platform_interface/lib/src/enums/enum_exports.dart
@@ -1,6 +1,5 @@
 // exports all enums
 
-export 'camera_source.dart';
 export 'network_quality_level.dart';
 export 'network_quality_verbosity.dart';
 export 'room_state.dart';