+++ programmable_video_platform_interface/lib/src/models/capturers/camera_event.dart
@@ -1,4 +1,3 @@
-import 'package:enum_to_string/enum_to_string.dart';
 import 'package:twilio_programmable_video_platform_interface/twilio_programmable_video_platform_interface.dart';
 
 abstract class BaseCameraEvent {
@@ -7,7 +6,7 @@ abstract class BaseCameraEvent {
   const BaseCameraEvent(this.model);
 
   @override
-  String toString() => 'CameraEvent: { cameraSource: ${EnumToString.convertToString(model!.source)} }';
+  String toString() => 'CameraEvent: { source: ${model!.source} }';
 }
 
 /// Use this event if camera was switched
@@ -17,7 +16,7 @@ class CameraSwitched extends BaseCameraEvent {
   ) : super(model);
 
   @override
-  String toString() => 'CameraSwitchedEvent: { cameraSource: ${EnumToString.convertToString(model!.source)} }';
+  String toString() => 'CameraSwitchedEvent: { source: ${model!.source} }';
 }
 
 /// Use this event if camera was switched
@@ -27,7 +26,7 @@ class FirstFrameAvailable extends BaseCameraEvent {
   ) : super(model);
 
   @override
-  String toString() => 'FirstFrameAvailableEvent: { cameraSource: ${EnumToString.convertToString(model!.source)} }';
+  String toString() => 'FirstFrameAvailableEvent: { source: ${model!.source} }';
 }
 
 /// Use this event if camera was switched
@@ -40,7 +39,7 @@ class CameraError extends BaseCameraEvent {
   ) : super(model);
 
   @override
-  String toString() => 'CameraErrorEvent: { cameraSource: ${EnumToString.convertToString(model!.source)}, exception: $exception }';
+  String toString() => 'CameraErrorEvent: { source: ${model!.source}, exception: $exception }';
 }
 
 class SkipableCameraEvent extends BaseCameraEvent {