+++ programmable_video_platform_interface/test/programmable_video_platform_interface_test.dart
@@ -2,6 +2,7 @@ import 'dart:typed_data';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
+import 'package:twilio_programmable_video_platform_interface/src/camera_source.dart';
 import 'package:twilio_programmable_video_platform_interface/src/method_channel_programmable_video.dart';
 import 'package:twilio_programmable_video_platform_interface/src/programmable_video_platform_interface.dart';
 
@@ -33,7 +34,8 @@ void main() {
       expect(() => ProgrammableVideoPlatform.instance.sendMessage('message', 'name'), throwsUnimplementedError);
       expect(() => ProgrammableVideoPlatform.instance.sendBuffer(Uint8List(0).buffer, 'name'), throwsUnimplementedError);
       expect(() => ProgrammableVideoPlatform.instance.enableAudioTrack(true, ''), throwsUnimplementedError);
-      expect(() => ProgrammableVideoPlatform.instance.switchCamera(), throwsUnimplementedError);
+      expect(() => ProgrammableVideoPlatform.instance.switchCamera(CameraSource('FRONT_CAMERA', false, false, false)), throwsUnimplementedError);
+      expect(() => ProgrammableVideoPlatform.instance.getSources(), throwsUnimplementedError);
       expect(() => ProgrammableVideoPlatform.instance.roomStream(0), throwsUnimplementedError);
       expect(() => ProgrammableVideoPlatform.instance.remoteParticipantStream(0), throwsUnimplementedError);
       expect(() => ProgrammableVideoPlatform.instance.localParticipantStream(0), throwsUnimplementedError);